pyproject.toml
[project]
name = "acapy-vc-authn-oidc"
version = "2.4.0"
description = "Verifiable Credential Identity Provider for OpenID Connect."
authors = [ { name = "OpenWallet Foundation" } ]
license = "Apache-2.0"
readme = "README.md"
classifiers = [
            "Programming Language :: Python :: 3",
            "License :: OSI Approved :: Apache Software License",
	        "Operating System :: OS Independent",
]
repository = "https://github.com/openwallet-foundation/acapy-vc-authn-oidc"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
fastapi = "^0.135.0"
jinja2 = "^3.1.6"
oic = "^1.7.0"
pymongo = "^4.15.4"
pyop = { version = "^3.4.2", extras = ["redis"] }
python-multipart = "^0.0.22" # required by fastapi to serve/upload files
qrcode = { version = "^8.2", extras = ["pil"]}
structlog = "^25.5.0"
uvicorn = { version = "^0.41.0", extras = ["standard"] }
redis = "^7.3.0"
canonicaljson = "^2.0.0" # used to provide unique consistent user identifiers"
pydantic-settings = "^2.12.0"
pyjwt = "^2.11.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.5"
mongomock = "^4.3.0"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
pytest = "^9.0.1"
respx = "^0.22"

[tool.pytest.ini_options]
testpaths = ["oidc-controller"]
pythonpath = ["oidc-controller"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
addopts = "-m 'not e2e and not integration'"
markers = [
    "integration: integration tests (no external services required)",
    "e2e: end-to-end tests (requires running stack + bootstrap)",
    "connection: tests requiring USE_CONNECTION_BASED_VERIFICATION=true",
    "oob: tests requiring USE_CONNECTION_BASED_VERIFICATION=false",
]

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/integration_tests/*",
    "*/conftest.py",
]

[tool.pyright]
pythonVersion = "3.13"

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"
